Organize modals and add jest test to card generation function
