--- src/index.test.js
@@ -1,5 +1,0 @@
-const { JSDOM } = require('jsdom');
-
-const dom = new JSDOM('<!DOCTYPE html><html><body></body></html>');
-global.window = dom.window;
-global.document = dom.window.document;
